# Store optimizer state in memory to improve `CmaEsSampler` performance