--- a/package/samplers/restart_cmaes/restart_cmaes.py
+++ b/package/samplers/restart_cmaes/restart_cmaes.py
@@ -71,7 +71,7 @@ class RestartCmaEsSampler(BaseSampler):
         self._inc_popsize = inc_popsize
         self._store_optimizer_state_in_storage = store_optimizer_state_in_storage
         if not self._store_optimizer_state_in_storage:
-            self._optimizer_metadata_by_trial: dict[int, dict[str, Any]] = {}
+            self._optimizer_states_by_trial: dict[int, dict[str, Any]] = {}
 
         if restart_strategy not in (
             "ipop",
@@ -146,27 +146,23 @@ class RestartCmaEsSampler(BaseSampler):
             latest_trial = completed_trials[-1]
 
             if self._store_optimizer_state_in_storage:
-                popsize = latest_trial.system_attrs.get(
-                    self._attr_keys.popsize(), self._initial_popsize
-                )
-                n_restarts = latest_trial.system_attrs.get(self._attr_keys.n_restarts(), 0)
-                n_restarts_with_large = latest_trial.system_attrs.get(
-                    self._attr_keys.n_restarts_with_large, 0
-                )
-                poptype = latest_trial.system_attrs.get(self._attr_keys.poptype, "small")
-                small_n_eval = latest_trial.system_attrs.get(self._attr_keys.small_n_eval, 0)
-                large_n_eval = latest_trial.system_attrs.get(self._attr_keys.large_n_eval, 0)
+                source = latest_trial.system_attrs
             else:
                 latest_trial_id = latest_trial._trial_id
-                latest_trial_attr = self._optimizer_metadata_by_trial.get(latest_trial_id, {})
-                popsize = latest_trial_attr.get(self._attr_keys.popsize(), self._initial_popsize)
-                n_restarts = latest_trial_attr.get(self._attr_keys.n_restarts(), 0)
-                n_restarts_with_large = latest_trial_attr.get(
-                    self._attr_keys.n_restarts_with_large, 0
-                )
-                poptype = latest_trial_attr.get(self._attr_keys.poptype, "small")
-                small_n_eval = latest_trial_attr.get(self._attr_keys.small_n_eval, 0)
-                large_n_eval = latest_trial_attr.get(self._attr_keys.large_n_eval, 0)
+                source = self._optimizer_states_by_trial.get(latest_trial_id, {})
+
+            attr_defaults = [
+                (self._attr_keys.popsize(), self._initial_popsize),
+                (self._attr_keys.n_restarts(), 0),
+                (self._attr_keys.n_restarts_with_large, 0),
+                (self._attr_keys.poptype, "small"),
+                (self._attr_keys.small_n_eval, 0),
+                (self._attr_keys.large_n_eval, 0),
+            ]
+
+            popsize, n_restarts, n_restarts_with_large, poptype, small_n_eval, large_n_eval = (
+                source.get(key, default) for key, default in attr_defaults
+            )
 
         optimizer = self._restore_optimizer(completed_trials, n_restarts)
         if optimizer is None:
@@ -191,7 +187,7 @@ class RestartCmaEsSampler(BaseSampler):
             completed_trials, optimizer.generation, n_restarts
         )
 
-        optimizer_metadata: dict[str, Any] = {}
+        optimizer_states: dict[str, Any] = {}
 
         if len(solution_trials) >= popsize:
             solutions: list[tuple[np.ndarray, float]] = []
@@ -247,7 +243,7 @@ class RestartCmaEsSampler(BaseSampler):
                         trial._trial_id, key, optimizer_attrs[key]
                     )
                 else:
-                    optimizer_metadata[key] = optimizer_attrs[key]
+                    optimizer_states[key] = optimizer_attrs[key]
 
         # Caution: optimizer should update its seed value.
         seed = self._cma_rng.rng.randint(1, 2**16) + trial.number
@@ -259,38 +255,26 @@ class RestartCmaEsSampler(BaseSampler):
                     trial._trial_id, "x_for_tell", x_for_tell.tolist()
                 )
             else:
-                optimizer_metadata["x_for_tell"] = x_for_tell.tolist()
+                optimizer_states["x_for_tell"] = x_for_tell.tolist()
         else:
             params = optimizer.ask()
 
+        attrs = [
+            (self._attr_keys.generation(n_restarts), optimizer.generation),
+            (self._attr_keys.popsize(), popsize),
+            (self._attr_keys.n_restarts(), n_restarts),
+            (self._attr_keys.n_restarts_with_large, n_restarts_with_large),
+            (self._attr_keys.poptype, poptype),
+            (self._attr_keys.small_n_eval, small_n_eval),
+            (self._attr_keys.large_n_eval, large_n_eval),
+        ]
+
         if self._store_optimizer_state_in_storage:
-            generation_attr_key = self._attr_keys.generation(n_restarts)
-            study._storage.set_trial_system_attr(
-                trial._trial_id, generation_attr_key, optimizer.generation
-            )
-            popsize_attr_key = self._attr_keys.popsize()
-            study._storage.set_trial_system_attr(trial._trial_id, popsize_attr_key, popsize)
-            n_restarts_attr_key = self._attr_keys.n_restarts()
-            study._storage.set_trial_system_attr(trial._trial_id, n_restarts_attr_key, n_restarts)
-            study._storage.set_trial_system_attr(trial._trial_id, self._attr_keys.poptype, poptype)
-            study._storage.set_trial_system_attr(
-                trial._trial_id, self._attr_keys.small_n_eval, small_n_eval
-            )
-            study._storage.set_trial_system_attr(
-                trial._trial_id, self._attr_keys.large_n_eval, large_n_eval
-            )
+            for key, value in attrs:
+                study._storage.set_trial_system_attr(trial._trial_id, key, value)
         else:
-            generation_attr_key = self._attr_keys.generation(n_restarts)
-            optimizer_metadata[generation_attr_key] = optimizer.generation
-            popsize_attr_key = self._attr_keys.popsize()
-            optimizer_metadata[popsize_attr_key] = popsize
-            n_restarts_attr_key = self._attr_keys.n_restarts()
-            optimizer_metadata[n_restarts_attr_key] = n_restarts
-            optimizer_metadata[self._attr_keys.n_restarts_with_large] = n_restarts_with_large
-            optimizer_metadata[self._attr_keys.poptype] = poptype
-            optimizer_metadata[self._attr_keys.small_n_eval] = small_n_eval
-            optimizer_metadata[self._attr_keys.large_n_eval] = large_n_eval
-            self._optimizer_metadata_by_trial[trial._trial_id] = optimizer_metadata
+            optimizer_states.update(attrs)
+            self._optimizer_states_by_trial[trial._trial_id] = optimizer_states
 
         external_values = trans.untransform(params)
 
@@ -360,7 +344,7 @@ class RestartCmaEsSampler(BaseSampler):
             items = (
                 trial.system_attrs.items()
                 if self._store_optimizer_state_in_storage
-                else self._optimizer_metadata_by_trial.get(trial._trial_id, {}).items()
+                else self._optimizer_states_by_trial.get(trial._trial_id, {}).items()
             )
             optimizer_attrs = {
                 key: value
@@ -457,18 +441,18 @@ class RestartCmaEsSampler(BaseSampler):
         self, trials: list[FrozenTrial], generation: int, n_restarts: int
     ) -> list[FrozenTrial]:
         generation_attr_key = self._attr_keys.generation(n_restarts)
-        solution_trials = []
-        for t in trials:
-            generation_of_trial = (
-                t.system_attrs.get(generation_attr_key, -1)
-                if self._store_optimizer_state_in_storage
-                else self._optimizer_metadata_by_trial.get(t._trial_id, {}).get(
+        if self._store_optimizer_state_in_storage:
+
+            def get_generation(t: FrozenTrial) -> int:
+                return t.system_attrs.get(generation_attr_key, -1)
+        else:
+
+            def get_generation(t: FrozenTrial) -> int:
+                return self._optimizer_states_by_trial.get(t._trial_id, {}).get(
                     generation_attr_key, -1
                 )
-            )
-            if generation_of_trial == generation:
-                solution_trials.append(t)
-        return solution_trials
+
+        return [t for t in trials if generation == get_generation(t)]
 
     def before_trial(self, study: optuna.Study, trial: FrozenTrial) -> None:
         self._independent_sampler.before_trial(study, trial)