## Contributor Agreements

Please read the [contributor agreements](https://github.com/optuna/optunahub-registry/blob/main/CONTRIBUTING.md#contributor-agreements) and if you agree, please click the checkbox below.

- [x] I agree to the contributor agreements.

> [!TIP]
> Please follow the [Quick TODO list](https://github.com/optuna/optunahub-registry/tree/main?tab=readme-ov-file#quick-todo-list-towards-contribution) to smoothly merge your PR.

## Motivation
The `CmaEsSampler` with `restart_strategy` requires a relatively large number of trials, but the current implementation stores the optimizer state in storage, which slows down execution and makes it difficult to execute a large number of trials.  
By allowing the optimizer state to be stored in memory instead, performance can be significantly improved.

## Description of the changes
- Add `store_optimizer_state_in_storage` option to choose whether the optimizer state is stored in memory or in storage.
- When `store_optimizer_state_in_storage=False`, the optimizer state is stored in in-memory dict, `self._optimizer_metadata_by_trial`.