@sawa3030 2025-05-16
This PR conflicts with [#266](https://github.com/optuna/optunahub-registry/pull/266), so it will be updated and made ready for review after #266 is merged.