@sawa3030 2025-05-16
I verified the changes using the original Optuna test cases for `CmaEsSampler`, excluding tests related to the removed arguments.  The test code is available [here](https://github.com/sawa3030/optunahub-registry/blob/tests-for-restart-cmaes/package/samplers/restart_cmaes/tests/test_cmaes.py)