@sawa3030 2025-05-16
Referring to [this comment](https://github.com/optuna/optuna/pull/6025#pullrequestreview-2780862086), I have confirmed that this PR behaves consistently with the original `CmaEsSampler` in Optuna v4.3.0. The code used for verification is shown below.

```
import optuna
import optunahub

from optuna.visualization import plot_optimization_history
from plotly.io import show
 
bbob = optunahub.load_module("benchmarks/bbob")
sphere2d = bbob.Problem(function_id=1, dimension=2, instance_id=1)

storage = optuna.storages.get_storage("sqlite:///cmaes_benchmark.db")

sampler = optuna.samplers.CmaEsSampler(seed=1, restart_strategy="ipop")
study_v4_3_0 = optuna.create_study(
    sampler=sampler,
    directions=sphere2d.directions,
    study_name="cmaes_v4_3_0",
    storage=storage
)
study_v4_3_0.optimize(sphere2d, n_trials=500)

module = optunahub.load_module(
    package="samplers/restart_cmaes",
    repo_owner="sawa3030",
    ref="fix/use-in-memory-in-restartcmaessampler",
)
RestartCmaEsSampler = module.RestartCmaEsSampler
sampler = RestartCmaEsSampler(seed=1, restart_strategy="ipop")
study_pr266 = optuna.create_study(
    sampler=sampler,
    directions=sphere2d.directions,
    study_name="cmaes_pr269",
    storage=storage
)
study_pr266.optimize(sphere2d, n_trials=500)

```

![image](https://github.com/user-attachments/assets/161661b3-71f5-41b8-9a83-f93560cc4ddc)