@sawa3030 2025-05-16
## Benchmark
The runtime of `sample_relative` was compared between this PR and the `main` branch of `optunahub-registry` using the following code.  
This PR reduces the overhead of `set_trial_system_attr`, resulting in a noticeable performance improvement.

```
import optuna
import optunahub
import cProfile

module = optunahub.load_module(
    package="samplers/restart_cmaes",
    repo_owner="sawa3030",
    ref="fix/use-in-memory-in-restartcmaessampler",
)
RestartCmaEsSampler = module.RestartCmaEsSampler

def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -1, 1)
    y = trial.suggest_int("y", -1, 1)
    return x**2 + y

sampler = RestartCmaEsSampler(restart_strategy="ipop")
study = optuna.create_study(sampler=sampler)
cProfile.run("study.optimize(objective, n_trials=1000)", filename="cmaes-in.stats")

```

### This PR
![image](https://github.com/user-attachments/assets/1b02c139-e91b-46ab-a8ce-abec3b9e8bc7)

### main
![image](https://github.com/user-attachments/assets/bbe5edfc-37f1-4820-8937-db243699e347)