@sawa3030 2025-06-18
The runtime of `study.optimize()` when using in-memory storage was compared between the master branch and this PR. The following script was executed alternately 10 times for each version, and the average and standard deviation of the runtimes were measured. Although the difference is small, the results suggest that this PR does not introduce any noticeable overhead even when using in-memory storage.

| Branch    | Average Runtime |
|-----------|------------------|
| master   | 1.620 ± 0.2146     | 
| this PR    | 1.524 ± 0.0867 |

```
import optuna
import optunahub

module = optunahub.load_module(
    package="samplers/restart_cmaes",
)
# module = optunahub.load_module(
#     package="samplers/restart_cmaes",
#     repo_owner="sawa3030",
#     ref="fix/use-in-memory-in-restartcmaessampler",
# )
RestartCmaEsSampler = module.RestartCmaEsSampler


def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -1, 1)
    y = trial.suggest_int("y", -1, 1)
    return x**2 + y


sampler = RestartCmaEsSampler(restart_strategy="ipop")
study = optuna.create_study(sampler=sampler)
study.optimize(objective, n_trials=200)
```