@gen740 2025-06-20
I ran the benchmark (provided by @sawa3030) using [hyperfine](https://github.com/sharkdp/hyperfine) and confirmed that this PR would likely improve the performance across all types of storage.

| Storage  | Master              | This PR             |
| -------- | ------------------- | ------------------- |
| InMemory | `Time (mean ± σ): 249.7 ms ±   3.1 ms`</br>`Range (min … max): 244.4 ms … 254.0 ms `</br>`11 runs` | `Time (mean ± σ): 246.3 ms ±   2.8 ms`</br>`Range (min … max): 242.0 ms … 249.6 ms `</br>`11 runs` |
| Journal  | `Time (mean ± σ): 688.8 ms ±  37.2 ms`</br>`Range (min … max): 620.1 ms … 733.1 ms `</br>`10 runs` | `Time (mean ± σ): 439.7 ms ±  15.3 ms`</br>`Range (min … max): 417.0 ms … 459.3 ms `</br>`10 runs` |
| SqLite   | `Time (mean ± σ):  3.233 s ±  0.080 s`</br>`Range (min … max):  3.148 s …  3.372 s `</br>`10 runs` | `Time (mean ± σ):  2.334 s ±  0.030 s`</br>`Range (min … max):  2.297 s …  2.390 s `</br>`10 runs` |
| MySQL    | `Time (mean ± σ):  9.795 s ±  2.859 s`</br>`Range (min … max):  7.944 s … 15.509 s `</br>`10 runs` | `Time (mean ± σ):  5.010 s ±  0.082 s`</br>`Range (min … max):  4.903 s …  5.154 s `</br>`10 runs` |