@gen740 2025-07-01
I think the user might be confused when to set the `store_optimizer_state_in_storage`. If this option is important to use in the multi-threaded case, I think it's better to say "This option should be set to `True` in the multi-threaded case".